+++ packages/mobile/layout/Spacer.tsx
@@ -1,6 +1,6 @@
 import React, { memo } from 'react';
 import { Animated, View, ViewProps, ViewStyle } from 'react-native';
-import { SharedProps } from '@cbhq/cds-common';
+import { FixedValue, SharedProps } from '@cbhq/cds-common';
 import { SpacerBaseProps } from '@cbhq/cds-common/types/SpacerBaseProps';
 import { getSpacerStyle } from '@cbhq/cds-common/utils/getSpacerStyle';
 
@@ -36,18 +36,20 @@ export const Spacer = memo(function Spacer({
     <Component
       {...viewProps}
       accessibilityRole="none"
-      style={getSpacerStyle({
-        flexGrow,
-        flexShrink,
-        flexBasis,
-        horizontal,
-        vertical,
-        maxHorizontal,
-        maxVertical,
-        minHorizontal,
-        minVertical,
-        spacingScaleValues: spacing,
-      })}
+      style={
+        getSpacerStyle({
+          flexGrow,
+          flexShrink,
+          flexBasis: flexBasis as FixedValue,
+          horizontal,
+          vertical,
+          maxHorizontal,
+          maxVertical,
+          minHorizontal,
+          minVertical,
+          spacingScaleValues: spacing,
+        }) as ViewStyle
+      }
     />
   );
 });